tests/wpt/mozilla/tests/mozilla/binding_keyword.html

<!doctype html>
<meta charset="utf-8">
<title></title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
  var t = new TestBinding();

  var dict = t.receiveTestDictionaryWithSuccessOnKeyword();
  assert_equals(dict.type, "success");

  var is_success = t.typeKeywordIsSuccess(dict);
  assert_true(is_success);
}, "Conversion of dictionary elements with rust keywords works")
</script>
